refactor(web): split WordBankEntryDetail into thin orchestrator per ADR 0006
The file had ballooned to ~530 lines with 5 components, a utility
function, and constants all in one file — the exact pattern ADR 0006
was written to prevent.

Extracted into:

  components/                     ← reusable across screens
    renderEmphasis.tsx            ← pure markdown emphasis utility
    ExamplesDisplay.tsx           ← parsed example list renderer

  screens/WordBankEntryDetail/    ← screen sub-components
    index.tsx                     ← thin orchestrator (~220 lines)
    ContentField.tsx              ← extracted definition/nuance/examples loop
    InlineEdit.tsx                ← generic inline text/textarea editing
    TagsRow.tsx                   ← tag chips + inline add input
    constants.ts                  ← EditableField type + FIELD_LABELS

Import path unchanged — App.tsx resolves to index.tsx automatically.
