+++ packages/web/src/screens/WordBankEntryDetail/constants.ts
@@ -1,0 +1,12 @@
+export type EditableField =
+  | "definition"
+  | "translationArabic"
+  | "nuance"
+  | "examples";
+
+export const FIELD_LABELS: Record<EditableField, string> = {
+  definition: "Definition",
+  translationArabic: "Translation",
+  nuance: "Nuance",
+  examples: "Examples",
+};
